Refactor bottom sheet handling in MainActivity and MovieGridItem
